refactor(todo-list): make components for each parts to split logic and better for maintenence, also add action to status checkbox to update database and ui
