+++ components/StickyNote/index.ts
@@ -1,0 +1,3 @@
+import TodoList from "./TodoList";
+
+export { TodoList };
