+++ components/index.ts
@@ -1,7 +1,9 @@
 import * as Form from "./Form";
 import * as Layout from "./Layout";
+import * as Filter from "./Filter";
+import * as StickyNote from "./StickyNote";
 
 import Modal from "./Modal";
 import * as Spinner from "./Spinner";
 
-export { Form, Layout, Modal, Spinner };
+export { Form, Layout, Filter, StickyNote, Modal, Spinner };
