+++ interfaces/frontend/filter.ts
@@ -1,0 +1,6 @@
+export type TStatusFilter = "all" | "incomplete" | "completed";
+
+export interface INameList {
+  _id: string;
+  displayName: string;
+}
