+++ interfaces/frontend/index.ts
@@ -1,0 +1,3 @@
+import * as IFilter from "./filter";
+
+export { IFilter };
