+++ utils/index.ts
@@ -1,0 +1,3 @@
+import * as converter from "./converter";
+
+export { converter };
